Small fixes
For pd.read_csv, added engine='python' because HyPAT was generating this warning: ParserWarning: Falling back to the 'python' engine because the 'c' engine does not support skipfooter; you can avoid this warning by specifying engine='python'.

Changed HyPAT so that loading data on a different tab doesn't cause the Export to Excel button to export the wrong data.