Updating math
Adjusted the minimum boundary for dt on AP and PP tab so that the additive time constant can be slightly negative. Also, note that, for the Ni data on the PP tab, it seems like making the "+1" change in the last commit lead to the curve fit falling into a local minimum and generating a worse fit for some files, but the theory is still correct. As such, we didn't revert the commit.

On the AP tab, adjusted the initial number of moles in the sample for pressures after the first to be the final number of moles in the previous pressure. (N/A to PP tab)

On the AP tab, adjusted the initial number of moles in the container to include the moles that are in the sample container after the isolation valve closes (N/A to PP tab)

Adjusted pseudo-data to account for the above changes.

Made the pseudo-sample twice as thick and adjusted parameters in the program to account for that. Made the absortion time substantially longer to cut back on warnings.